+++ feature/main/info/series/src/main/java/com/jparkbro/info/series/InfoSeriesAction.kt
@@ -3,6 +3,6 @@ package com.jparkbro.info.series
 interface InfoSeriesAction {
     data object OnRetryClicked : InfoSeriesAction
     data object NavigateBack : InfoSeriesAction
-    data class NavigateToInfoAnime(val animeId: Int) : InfoSeriesAction
+    data class NavigateToInfoAnime(val animeId: Long) : InfoSeriesAction
     data object LoadMoreAnime : InfoSeriesAction
 }
